requirements.txt
langchain==0.3.25
langchain-chroma==0.2.3
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
openai==1.76.2
python-dotenv==1.1.0
chromadb==0.6.3
pandas==2.2.3
numpy==2.2.5
pydantic==2.11.4
fastapi==0.115.9
uvicorn==0.34.2
